Extract XML message parsing logic
